filename: renamed monthly-client-projects-os-project-list to monthly-client-projects-os-projects
